Change version type from long to int

+++ Source/Streamstone.Tests/Api.cs
@@ -8,7 +8,7 @@
         public const string EventRowKeyPrefix = "SS-SE-";
         public const string EventIdRowKeyPrefix = "SS-UID-";
 
-        public static string FormatEventRowKey(this long version) => $"{EventRowKeyPrefix}{version:d10}";
+        public static string FormatEventRowKey(this int version) => $"{EventRowKeyPrefix}{version:d10}";
         public static string FormatEventIdRowKey(this string id) => $"{EventIdRowKeyPrefix}{id}";
     }
 }
