+++ Source/Streamstone.Tests/Scenarios/Writing_to_stream.cs
@@ -95,8 +95,8 @@ namespace Streamstone.Scenarios
             EventData[] events = {CreateEvent("e1"), CreateEvent("e2")};
             var result = await Stream.WriteAsync(stream, events);
 
-            AssertModifiedStream(stream, result, version: 2L);
-            AssertStreamEntity(version: 2L);
+            AssertModifiedStream(stream, result, version: 2);
+            AssertStreamEntity(version: 2);
 
             var storedEvents = result.Events;
             Assert.That(storedEvents.Length, Is.EqualTo(2));
@@ -324,7 +324,7 @@ namespace Streamstone.Scenarios
             AssertStreamEntity(3, replaced);
         }
 
-        void AssertNewStream(StreamWriteResult actual, long version, object properties = null)
+        void AssertNewStream(StreamWriteResult actual, int version, object properties = null)
         {
             var newStream = actual.Stream;
             var newStreamEntity = partition.RetrieveStreamEntity();
@@ -333,7 +333,7 @@ namespace Streamstone.Scenarios
             expectedStream.ToExpectedObject().ShouldEqual(newStream);
         }
 
-        void AssertModifiedStream(Stream previous, StreamWriteResult actual, long version)
+        void AssertModifiedStream(Stream previous, StreamWriteResult actual, int version)
         {
             var actualStream = actual.Stream;
             var actualStreamEntity = partition.RetrieveStreamEntity();
@@ -344,7 +344,7 @@ namespace Streamstone.Scenarios
             expectedStream.ToExpectedObject().ShouldEqual(actualStream);
         }
 
-        Stream CreateStream(long version, ETag etag, object properties = null)
+        Stream CreateStream(int version, ETag etag, object properties = null)
         {
             var props = properties != null  
                 ? StreamProperties.From(properties) 
@@ -353,7 +353,7 @@ namespace Streamstone.Scenarios
             return new Stream(partition, etag, version, props);
         }
 
-        void AssertStreamEntity(long version = 0, object properties = null)
+        void AssertStreamEntity(int version = 0, object properties = null)
         {
             var newStreamEntity = partition.RetrieveStreamEntity();
 
@@ -369,13 +369,13 @@ namespace Streamstone.Scenarios
             expectedEntity.ToExpectedObject().ShouldMatch(newStreamEntity);
         }
 
-        void AssertRecordedEvent(long version, EventData source, RecordedEvent actual)
+        void AssertRecordedEvent(int version, EventData source, RecordedEvent actual)
         {
             var expected = source.Record(partition, version);
             expected.ToExpectedObject().ShouldMatch(actual);
         }
 
-        static void AssertEventEntity(long version, EventEntity actual)
+        static void AssertEventEntity(int version, EventEntity actual)
         {
             var expected = new
             {
@@ -391,7 +391,7 @@ namespace Streamstone.Scenarios
             expected.ToExpectedObject().ShouldMatch(actual);
         }
 
-        static void AssertEventIdEntity(string id, long version, EventIdEntity actual)
+        static void AssertEventIdEntity(string id, int version, EventIdEntity actual)
         {
             var expected = new
             {
